Add concordance index function
This commit includes a function for calculating Harrell's concordance
index, which can be calculated in R using 'hmisc'. The function is
implemented in Fortran, with a small Python wrapper. The reason for this
is that since calculating the C-index is an O(n^2) process, it quickly
becomes unacceptably slow with pure Python. Comparing a pure Python
implementation with the Fortran version, on arrays with length 1000,
Python required 434 ms while Fortran did it in 4.73 ms. So almost a
factor of 100 difference.

As a consequence of the addition of the Fortran module, the setup
script now utilizes numpy's setup function which will handle the
compilation of the native code.

In addition, a small unit test has been added. To be able to run the
unit tests, it is likely necessary to compile the native code first with:

    python setup.py build_ext --inplace

Signed-off-by: Jonas Kalderstam <jonas@kalderstam.se>