Include tests in final package
It's nice to be able to run the tests, even as an end-user.  Tests were
not included earlier at least on Windows when installing from the wheel
packages.

This also greatly simplifies Windows testing capabilities. It will make
it possible to automate tests where the wheel packages are installed and
the tests executed.

Signed-off-by: Jonas Kalderstam <jonas@kalderstam.se>