Make lifelines work in pure python mode
Added a pure python version of concordance index,
as a fallback option, for developers without access
to a proper compiler (Windows people). This allows for
easier contribution to the library, while still allowing
us to maintain the higher speed for users who install from
pypi.

The setup script will now try to compile the extensions, and if
it fails, it will do it again with just the python source.

Signed-off-by: Jonas Kalderstam <jonas@kalderstam.se>