Adds the SBG package to lifelines.
It includes the high level object that interact with pandas as well as
two lower level scripts that handled the calculations.

Additionally it also includes a example notebook and a data generating script.

The data generating script could, and will be, integrated with the existing generate_datasets.py
script. And the notebook must be revised to match the new API.