Added plot and minor tweaks
Verified the plot functionality with AalenJohansenFitter. Added to the documentation. Cleaned up the code a bit. 

For the future, might consider adding the Aalen confidence interval method. Currently uses the Delta method. Not a big difference between the two. Both are in the SAS document I linked in the _bounds docs