ParametricRegressionFitter: Use functools.partial
For _create_neg_likelihood_with_penalty_function()
use the functools.partial approach instead of function inside
function so that it can be pickled correctly.

Also, save the _initial_point_dict so that the unflatten function
can be generated on the fly and the autograd function does not need
to be a property that needs to be saved.